+++ CHANGELOG.md
@@ -13,7 +13,7 @@ All notable changes to this project will be documented in this file.
 - mosaic layout: clamp ratio to 32/9
 - Video: disable subtitles by default
 - Map: Stamen Watercolor layer (no longer served for free by Stamen) now served by Smithsonian Institution
-- upgraded Flutter to stable v3.13.6
+- upgraded Flutter to stable v3.13.7
 
 ### Removed